Removed Record.spatialObject(). It is not reasonable to assume that indexes always contain the spatial object. E.g., in a relational database system, the index will contain the z-value and the tables's PK, and the referenced row will contain the spatial object.
Because of this change, application of the SpatialJoin must be
optional, a spatial join can now be run without a SpatialJoinFilter.
Also, the spatial object must be passed to SpatialIndex.add, it can no
longer be obtained form the Record.

The SpatialJoin API is changed. Instead of
SpatialJoin.newSpatialJoin(filter, duplicates).iterator(leftIndex,
rightIndex), everything is done in the iterator call:
SpatialJoin.iterator(leftIndex, rightIndex, filter, duplicates).

SpatialJoinFilter has been generalized (and genericized) to allow for
comparison of two records (many/many join) and a spatial object and a
record (one/many join).